working

+++ angular/src/app/components/audit-history/audit-history.component.ts
@@ -1,3 +1,4 @@
+import { ConfigStateService } from '@abp/ng.core';
 import { NgFor } from '@angular/common';
 import { Component, Input, OnInit } from '@angular/core';
 import {
@@ -36,7 +37,7 @@ export class AuditHistoryComponent implements OnInit {
       actionItem: new FormControl(''),
     }),
   });
-  unauthorizedPerson: boolean = true;
+  unauthorizedPerson: boolean = false;
   displayedColumns = [
     'Date Of Audit',
     'Reviewed By',
@@ -49,9 +50,10 @@ export class AuditHistoryComponent implements OnInit {
   constructor(
     private fb: FormBuilder,
     private route: ActivatedRoute,
-    private auditHistoryService: AuditHistoryService
+    private auditHistoryService: AuditHistoryService,
+    private config:ConfigStateService
   ) {
-    this.projectId = this.route.snapshot.pathFromRoot[1].params['id'];
+    this.projectId = this.route.snapshot.pathFromRoot[2].params['id'];
   }
 
   ngOnInit(): void {
@@ -64,14 +66,18 @@ export class AuditHistoryComponent implements OnInit {
       error => {
         this.addExistingData([]);
         if (error.status == 403) {
-          this.unauthorizedPerson = false;
-          console.log(this.unauthorizedPerson);
           console.warn('Unauthorized access (403):', error);
         } else {
           console.error('Error fetching projects:', error);
         }
       }
     );
+    if (
+      this.config.getOne('currentUser').roles[0] == 'client' ||
+      this.config.getOne('currentUser').roles[0] == 'projectManager'
+    ) {
+      this.unauthorizedPerson = true;
+    }
   }
 
   addExistingData(data: any): void {
@@ -119,6 +125,10 @@ export class AuditHistoryComponent implements OnInit {
   }
 
   removeRow(index: number): void {
+    const getConfirmation=window.confirm("Do you want to delte");
+    if(getConfirmation==false){
+      return;
+    }
     const approveteamArray = this.forms.get('formitem') as FormArray;
     const controlAtIndex = approveteamArray.at(index);
     this.auditHistoryService.deleteItem(controlAtIndex.value.id).subscribe(
@@ -133,7 +143,7 @@ export class AuditHistoryComponent implements OnInit {
   }
 
   onSubmit(): void {
-    if (this.forms.valid) {
+    if (this.forms.valid && this.unauthorizedPerson==false) {
       this.forms.value.formitem.forEach(async e => {
         try {
           const modelDate: AuditHistoryModel = {
@@ -157,9 +167,8 @@ export class AuditHistoryComponent implements OnInit {
           console.error('Error:', error);
         }
       });
-    }
-    else{
-      alert("Make sure you filled right value, check date formate");
+    } else {
+      alert('Make sure you filled right value, check date formate');
     }
   }
 }
