+++ angular/src/app/components/escalation-matrix/escalation-matrix.component.ts
@@ -1,3 +1,4 @@
+import { ConfigStateService } from '@abp/ng.core';
 import { NgFor } from '@angular/common';
 import { Component, Input, OnInit } from '@angular/core';
 import {
@@ -31,16 +32,17 @@ export class EscalationMatrixComponent implements OnInit {
       comments: new FormControl(''),
     }),
   });
-  unauthorizedPerson: boolean = true;
+  unauthorizedPerson: boolean = false;
   displayedColumns = ['Level', 'EscalationType'];
   levelType = ['Level1', 'Level2', 'Level3', 'Level4', 'Level5'];
   escalationType = ['Operational', 'Financial', 'Technical'];
   constructor(
     private fb: FormBuilder,
     private route: ActivatedRoute,
-    private escalationMatrixService: EscalationMatrixService
+    private escalationMatrixService: EscalationMatrixService,
+    private config: ConfigStateService
   ) {
-    this.projectId = this.route.snapshot.pathFromRoot[1].params['id'];
+    this.projectId = this.route.snapshot.pathFromRoot[2].params['id'];
   }
 
   ngOnInit(): void {
@@ -53,14 +55,18 @@ export class EscalationMatrixComponent implements OnInit {
       error => {
         this.addExistingData([]);
         if (error.status == 403) {
-          this.unauthorizedPerson = false;
-          console.log(this.unauthorizedPerson);
           console.warn('Unauthorized access (403):', error);
         } else {
           console.error('Error fetching projects:', error);
         }
       }
     );
+    if (
+      this.config.getOne('currentUser').roles[0] == 'client' ||
+      this.config.getOne('currentUser').roles[0] == 'auditor'
+    ) {
+      this.unauthorizedPerson = true;
+    }
   }
 
   addExistingData(data: any): void {
@@ -100,6 +106,10 @@ export class EscalationMatrixComponent implements OnInit {
   }
 
   removeRow(index: number): void {
+    const getConfirmation = window.confirm('Do you want to delte');
+    if (getConfirmation == false) {
+      return;
+    }
     const approveteamArray = this.forms.get('formitem') as FormArray;
     const controlAtIndex = approveteamArray.at(index);
     this.escalationMatrixService.deleteItem(controlAtIndex.value.id).subscribe(
@@ -114,7 +124,7 @@ export class EscalationMatrixComponent implements OnInit {
   }
 
   onSubmit(): void {
-    if (this.forms.valid) {
+    if (this.forms.valid && this.unauthorizedPerson==false) {
       this.forms.value.formitem.forEach(async e => {
         try {
           const modelDate: EscalationMatrixModel = {
